Service Layer, component registration

+++ src/main/java/com/gabriedevbr/portfolioWeb/entities/User.java
@@ -19,7 +19,7 @@ private static final long serialVersionUID = 1L;
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	private String email;
 	private Integer phone;
@@ -29,7 +29,7 @@ private static final long serialVersionUID = 1L;
 
 	}
 
-	public User(Integer id, String name, String email, Integer phone, Integer password) {
+	public User(Long id, String name, String email, Integer phone, Integer password) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -38,11 +38,11 @@ private static final long serialVersionUID = 1L;
 		this.password = password;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
